feat(dicom): switch upload metadata extraction to dcmjs
- UploadPage.parseDicomFile now uses extractAllDicomMetadataFromArrayBuffer
  (dcmjs parse, dicom-parser fallback) and the shared
  flattenDicomMetadataForGridFS helper instead of an inline flatten
- radiology-workflow TechDashboard (untracked package, edited in place)
  does the same — and now flattens before upload, fixing a latent bug
  where it POSTed the nested shape so /api/dicom/upload never persisted
  real UIDs into dicom.files metadata
- flatten helper covered by the node --test suite (8 passing)
